Fix  task, add logging in task

+++ rpihelper/transmission/tasks.py
@@ -3,9 +3,10 @@
 import os
 import os.path
 
-import transmissionrpc
-
 from rpihelper.celery import current_app, celery
+from rpihelper.transmission.logic import (
+    transmissionrpc_client, transmissionrpc_add_torrent,
+)
 
 __all__ = (
     'check_torrent_files',
@@ -14,21 +15,18 @@ __all__ = (
 
 @celery.task
 def check_torrent_files():
-    print('check_torrent_files')
-
-    try:
-        tc = transmissionrpc.Client(
-            address=current_app.config['TRANSMISSION_ADDRESS'],
-            port=current_app.config['TRANSMISSION_PORT'],
-            timeout=current_app.config['TRANSMISSION_TIMEOUT']
-        )
-    except transmissionrpc.error.TransmissionError as e:
-        # TODO: logging?
+    tc = transmissionrpc_client()
+    if not tc:
+        current_app.logger.info('No connetion to remote transmission, stop task.')
         return
 
-    transmission_directory = current_app.config['TRANSMISSION_DROPBOX_DIRECORY']
-    for dirpath, diranames, filenames in os.walk(transmission_directory):
+    walk_directory = current_app.config['TRANSMISSION_DROPBOX_DIRECTORY']
+    for dirpath, diranames, filenames in os.walk(walk_directory):
         for filename in filenames:
             file_path = os.path.join(dirpath, filename)
-            tc.add_torrent(file_path)  # TODO: try/except
-            os.remove(file_path)
+            success = transmissionrpc_add_torrent(tc, file_path)
+            if success:
+                os.remove(file_path)
+                current_app.logger.info('Successfully added torrent file "%s".' % file_path)
+            else:
+                current_app.logger.info('Torrent file "%s" not added, skip it.' % file_path)
